src/main/java/com/github/brunomndantas/jscrapper/support/parser/single/text/reference/SingleReferenceFloatTextParser.java
package com.github.brunomndantas.jscrapper.support.parser.single.text.reference;

import com.github.brunomndantas.jscrapper.support.parser.single.text.SingleTextParser;
import org.openqa.selenium.WebElement;

public class SingleReferenceFloatTextParser extends SingleTextParser {

    public SingleReferenceFloatTextParser() {
        super(null);
    }



    @Override
    protected Object parseElement(WebElement element) throws Exception {
        return Float.parseFloat(element.getText());
    }

}